show section only when data products are present
