+++ context/app/static/js/pages/Organ/Organ.tsx
@@ -10,7 +10,7 @@ import { OrganFile } from 'js/components/organ/types';
 import DetailLayout from 'js/components/detailPage/DetailLayout';
 import CellPopulationPlot from 'js/components/organ/CellPop';
 import DataProducts from 'js/components/organ/DataProducts';
-import { useAssayBucketsQuery, useHasSamplesQuery, useLabelledDatasetsQuery } from './hooks';
+import { useAssayBucketsQuery, useDataProducts, useHasSamplesQuery, useLabelledDatasetsQuery } from './hooks';
 
 interface OrganProps {
   organ: OrganFile;
@@ -33,6 +33,7 @@ function Organ({ organ }: OrganProps) {
   const assayBuckets = useAssayBucketsQuery(searchItems);
   const samplesHits = useHasSamplesQuery(searchItems);
   const labeledDatasetUuids = useLabelledDatasetsQuery(searchItems);
+  const dataProducts = useDataProducts(organ.name);
 
   const shouldDisplaySection: Record<string, boolean> = {
     [summaryId]: Boolean(organ?.description),
@@ -40,7 +41,7 @@ function Organ({ organ }: OrganProps) {
     [cellpopId]: labeledDatasetUuids.length > 0,
     [referenceId]: Boolean(organ?.azimuth),
     [assaysId]: assayBuckets.length > 0,
-    [dataProductsId]: true,
+    [dataProductsId]: dataProducts.length > 0,
     [samplesId]: samplesHits.length > 0,
   };
 
@@ -70,7 +71,11 @@ function Organ({ organ }: OrganProps) {
         bucketData={assayBuckets}
         shouldDisplay={shouldDisplaySection[assaysId]}
       />
-      <DataProducts id={dataProductsId} organName={organ.name} />
+      <DataProducts
+        id={dataProductsId}
+        dataProducts={dataProducts}
+        shouldDisplay={shouldDisplaySection[dataProductsId]}
+      />
       <Samples id={samplesId} organTerms={searchItems} shouldDisplay={shouldDisplaySection[samplesId]} />
     </DetailLayout>
   );
