+++ context/app/static/js/pages/Organ/hooks.ts
@@ -1,5 +1,9 @@
 import { useMemo } from 'react';
+import useSWR from 'swr';
+
 import useSearchData, { useSearchHits } from 'js/hooks/useSearchData';
+import { useAppContext } from 'js/components/Contexts';
+import { OrganDataProducts } from 'js/components/organ/types';
 import { mustHaveOrganClause } from './queries';
 
 export function useHasSamplesQuery(searchItems: string[]) {
@@ -152,4 +156,37 @@ export function useLabelledDatasetsQuery(searchItems: string[]) {
   );
   const { searchHits: datasetsHits } = useSearchHits<never>(datasetsQuery, { useDefaultQuery: false });
   return datasetsHits.map((hit) => hit._id);
+}
+
+export function useDataProducts(organName: string) {
+  const { dataProductsEndpoint } = useAppContext();
+  const dataProductsUrl = `${dataProductsEndpoint}/api/data_products/${organName}`;
+
+  const fetcher = async ({
+    url,
+    requestInit,
+  }: {
+    url: string;
+    requestInit: RequestInit;
+  }): Promise<OrganDataProducts[]> => {
+    const response = await fetch(url, requestInit);
+    if (!response.ok) throw new Error('Failed to fetch data products');
+    return response.json() as Promise<OrganDataProducts[]>;
+  };
+
+  const { data } = useSWR<OrganDataProducts[]>(dataProductsUrl, (url: string) =>
+    fetcher({
+      url,
+      requestInit: { headers: { 'Content-type': 'application/json', Accept: 'application/json' } },
+    }),
+  );
+
+  const dataProducts = data ?? [];
+
+  const dataProductsWithUUIDs = dataProducts.map((product) => {
+    const datasetUUIDs = product.dataSets.map((dataset) => dataset.uuid);
+    return { ...product, datasetUUIDs };
+  });
+
+  return dataProductsWithUUIDs;
 }
